+++ CHANGELOG.rst
@@ -15,3 +15,5 @@ first official Alpha release.
  * Added DeploySite model
  * Added Deploy plugin
  * Added jquery.cicle in media. jQuery plugin to handle rotating pictures.
+ * Added plugin's url for handling dynamic plugins not connected to a model.
+ * Added Contact form plugin.